+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Akeeba Panopticon Connector for Joomla 4.x and 5.x version 1.0.2
+
+* ✨ Collect server information
+
 # Akeeba Panopticon Connector for Joomla 4.x and 5.x version 1.0.1
 
 * 🐞 Linking Panopticon to Akeeba Backup may fail if the JSON API isn't already active